+++ test/benchmark/reasoner/iam/complex/LanguageFeatureTest.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.vaticle.typedb.core.reasoner.benchmark.iam;
+package com.vaticle.typedb.core.reasoner.benchmark.iam.complex;
 
 import com.vaticle.typedb.core.reasoner.benchmark.iam.common.Benchmark;
 import com.vaticle.typedb.core.reasoner.benchmark.iam.common.BenchmarkRunner;
@@ -27,23 +27,28 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
-public class ComplexRuleGraphTest {
+public class LanguageFeatureTest {
+    static final Path RESOURCE_DIRECTORY =  Paths.get("test", "benchmark", "reasoner", "iam", "complex");
+    private static final Path COMMON_RESOURCE_DIR = Paths.get("test", "benchmark", "reasoner", "iam", "resources");
 
-    private static final String database = "iam-benchmark-rules";
+    private static final String database = "iam-benchmark-language-features";
     private static final BenchmarkRunner benchmarker = new BenchmarkRunner(database);
     private final QueryParams queryParams;
 
-    public ComplexRuleGraphTest() {
-        queryParams = QueryParams.load();
+    public LanguageFeatureTest() {
+        queryParams = QueryParams.load(COMMON_RESOURCE_DIR.resolve("params.yml"));
     }
 
     @BeforeClass
     public static void setup() throws IOException {
         benchmarker.setUp();
-        benchmarker.loadSchema("schema_types.tql");
-        benchmarker.loadSchema("schema_rules_naive.tql");
-        benchmarker.importData("data.typedb");
+        benchmarker.loadSchema(COMMON_RESOURCE_DIR.resolve("types.tql"));
+        benchmarker.loadSchema(COMMON_RESOURCE_DIR.resolve("rules.tql"));
+        benchmarker.loadSchema(RESOURCE_DIRECTORY.resolve("language-features-test.tql"));
+        benchmarker.importData(COMMON_RESOURCE_DIR.resolve("data.typedb"));
         benchmarker.warmUp();
     }
 
@@ -58,37 +63,35 @@ public class ComplexRuleGraphTest {
     }
 
     @Test
-    public void testCombinatorialProofsSingle() {
+    public void testValuePredicateFiltering() {
         String query = String.format(
                 "match\n" +
-                        "$p isa person, has email \"%s\";\n" +
-                        "$f isa file, has path \"%s\";\n" +
-                        "$o isa operation, has name \"%s\";\n" +
+                        "$p isa person, has $email; $email = \"%s\";\n" +
+                        "$f isa file, has $path; $path = \"%s\";\n" +
+                        "$o isa operation, has $operation; $operation = \"%s\";\n" +
                         "$a (object: $f, action: $o) isa access;\n" +
-                        "$pe (subject: $p, access: $a) isa permission;\n",
+                        "$pe (subject: $p, access: $a) isa permission, has validity true;\n",
                 queryParams.permissionEmail, queryParams.permissionObject, queryParams.permissionAction);
-        Benchmark benchmark = new Benchmark("combinatorial-proofs-single", query, 1);
+        Benchmark benchmark = new Benchmark("value-predicate-filtering", query, 1);
         benchmarker.runBenchmark(benchmark);
+
         benchmark.assertAnswerCountCorrect();
-        benchmark.assertRunningTime(1000);
-        benchmark.assertCounters(200, 149, 301, 1658);
+        benchmark.assertRunningTime(2500);
+        benchmark.assertCounters(1000, 200, 500, 1500);
     }
 
     @Test
-    public void testCombinatorialProofsAll() {
+    public void variabilisedRules() {
         String query = String.format(
                 "match\n" +
                         "$p isa person, has email \"%s\";\n" +
-                        "$o isa object, has id $o-id;\n" +
-                        "$a isa action, has name $an;\n" +
-                        "$ac (object: $o, action: $a) isa access;\n" +
-                        "$pe (subject: $p, access: $ac) isa permission;\n",
+                        "(group: $g, member: $p) isa variabilised-group-membership;\n",
                 queryParams.permissionEmail);
-
-        Benchmark benchmark = new Benchmark("combinatorial-proofs-all", query, 67);
+        Benchmark benchmark = new Benchmark("variabilised-rules", query, 3);
         benchmarker.runBenchmark(benchmark);
+
         benchmark.assertAnswerCountCorrect();
         benchmark.assertRunningTime(1000);
-        benchmark.assertCounters(100, 265, 342, 837);
+        benchmark.assertCounters(200, 9, 29, 104);
     }
 }
