+++ test/benchmark/reasoner/iam/complex/conjunction-structure-test.tql
@@ -1,0 +1,34 @@
+#
+# Copyright (C) 2022 Vaticle
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+define
+
+high-arity-test-segregation-violation sub relation,
+    relates subject, relates object, relates action;
+subject plays high-arity-test-segregation-violation:subject;
+object plays high-arity-test-segregation-violation:object;
+action plays high-arity-test-segregation-violation:action;
+
+rule high-arity-segregation-violation-detection:
+when {
+    $ac1(object: $o, action: $a1) isa access;
+    $ac2(object: $o, action: $a2) isa access;
+    $p1(subject: $s, access: $ac1) isa permission;
+    $p2(subject: $s, access: $ac2) isa permission;
+} then {
+    (subject: $s, object: $o, action: $a1, action: $a2) isa high-arity-test-segregation-violation;
+};
