+++ test/benchmark/reasoner/iam/complex/language-features-test.tql
@@ -17,23 +17,6 @@
 
 # Ensure none of the new rules are triggered by only creating rules which conclude test-specific types
 define
-
-high-arity-test-segregation-violation sub relation,
-    relates subject, relates object, relates action;
-subject plays high-arity-test-segregation-violation:subject;
-object plays high-arity-test-segregation-violation:object;
-action plays high-arity-test-segregation-violation:action;
-
-rule high-arity-segregation-violation-detection:
-when {
-    $ac1(object: $o, action: $a1) isa access;
-    $ac2(object: $o, action: $a2) isa access;
-    $p1(subject: $s, access: $ac1) isa permission;
-    $p2(subject: $s, access: $ac2) isa permission;
-} then {
-    (subject: $s, object: $o, action: $a1, action: $a2) isa high-arity-test-segregation-violation;
-};
-
 variabilised-membership sub relation, relates parent, relates member; # abstract
 variabilised-collection-membership sub variabilised-membership, relates collection as parent;
 variabilised-set-membership sub variabilised-membership, relates set as parent;
