+++ test/benchmark/reasoner/iam/resources/large-data.typedb
